adjust attack frequency (time interval) to every 5 seconds
the malicious node was sending DIS bursts every 50ms, resulting in excessive logging output, this caused the node to sometimes miss logging some messages.

I also did this to make the simulation doable on my 'not very good' laptop (since sending bursts every 50ms consumes a lot of ressources, causing my laptop to lag). The main goal of this attack is to drain the network's energy by periodically flooding it with DIS messages, sending a burst every 5 seconds is sufficient to achieve this.